Open `ipynb` notebooks with UTF-8 encoding
Previously, some (if not all notebooks) would be read as windows1252
encoding (presumably only on windows), which read unicode symbols such
as zero-width spaces and fancy quotation marks as gobbeldygook. The
script then resaves this crap as UTF-8. If we happen to run the script
multiple times, the cursed text self-replicates, growing exponentially.

Consistently reading and writing the notebook files in UTF-8 eliminates
this issue.